+++ app/src/main/java/pt/rvcoding/cvnotes/domain/use_case/section/HasSelectedSections.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.flow.Flow
 import pt.rvcoding.cvnotes.data.SPKey
 import pt.rvcoding.cvnotes.domain.repository.SectionRepository
 import pt.rvcoding.cvnotes.domain.repository.SharedPreferencesRepository
+import pt.rvcoding.cvnotes.domain.util.sha256
 
 class HasSelectedSections(
     private val spRepository: SharedPreferencesRepository,