+++ app/src/main/java/pt/rvcoding/cvnotes/domain/util/Encoding.kt
@@ -1,0 +1,7 @@
+package pt.rvcoding.cvnotes.domain.util
+
+import android.util.Base64
+
+
+fun String.sha256(): String = Base64.encodeToString(this.toByteArray(), Base64.NO_WRAP)
+