+++ app/src/main/java/pt/rvcoding/cvnotes/ui/about/AboutState.kt
@@ -13,4 +13,6 @@ data class AboutState(
 
 data class AboutProfileState(
     val email: String = "",
+    val profession: String = "",
+    val professionOverride: Boolean = false
 )